Simplify Image class to pure NumPy, remove CUDA backend
- Rewrite Image.py as a simple dataclass with NumPy-only data
- Delete image_impls/ directory (AbstractImage, CudaImage, NumpyImage)
- Remove unused methods (solve_pnp, csrt_tracker, from_depth, to_depth_meters)
- Update all imports across codebase
- Simplify rectify_image to CPU-only path
