+++ dimos/mapping/occupancy/visualizations.py
@@ -21,8 +21,7 @@ from numpy.typing import NDArray
 
 from dimos.msgs.nav_msgs import Path
 from dimos.msgs.nav_msgs.OccupancyGrid import OccupancyGrid
-from dimos.msgs.sensor_msgs import Image
-from dimos.msgs.sensor_msgs.image_impls.AbstractImage import ImageFormat
+from dimos.msgs.sensor_msgs.Image import Image, ImageFormat
 
 Palette: TypeAlias = Literal["rainbow", "turbo"]
