+++ dimos/mapping/occupancy/visualize_path.py
@@ -18,8 +18,7 @@ import numpy as np
 from dimos.mapping.occupancy.visualizations import visualize_occupancy_grid
 from dimos.msgs.nav_msgs import Path
 from dimos.msgs.nav_msgs.OccupancyGrid import OccupancyGrid
-from dimos.msgs.sensor_msgs.Image import Image
-from dimos.msgs.sensor_msgs.image_impls.AbstractImage import ImageFormat
+from dimos.msgs.sensor_msgs.Image import Image, ImageFormat
 
 
 def visualize_path(
