+++ dimos/perception/common/__init__.py
@@ -34,8 +34,6 @@ from .utils import (
     project_3d_points_to_2d_cpu,
     project_3d_points_to_2d_cuda,
     rectify_image,
-    rectify_image_cpu,
-    rectify_image_cuda,
     setup_logger,
     torch,
     yaml,
@@ -77,8 +75,6 @@ __all__ = [
     "project_3d_points_to_2d_cpu",
     "project_3d_points_to_2d_cuda",
     "rectify_image",
-    "rectify_image_cpu",
-    "rectify_image_cuda",
     "setup_logger",
     "torch",
     "yaml",
