+++ dimos/protocol/pubsub/impl/jpeg_shm.py
@@ -16,8 +16,7 @@ from typing import Any
 
 from turbojpeg import TurboJPEG  # type: ignore[import-untyped]
 
-from dimos.msgs.sensor_msgs.Image import Image
-from dimos.msgs.sensor_msgs.image_impls.AbstractImage import ImageFormat
+from dimos.msgs.sensor_msgs.Image import Image, ImageFormat
 from dimos.protocol.pubsub.encoders import PubSubEncoderMixin
 from dimos.protocol.pubsub.impl.shmpubsub import SharedMemoryPubSubBase
